[refactor] : png 파일 svg 로 교체
- 화질개선을 위해 png 파일을 svg 파일로 교체함
- 책 등록 로직에서의 책장 이미지를 3d책장이미지로 변경함
- 사용되지 않는 png 파일 삭제 및 svg 파일 추가
